git-svn-id: https://taskunifier.svn.sourceforge.net/svnroot/taskunifier/TaskUnifierGui@218 4cc7d151-2d3f-4153-82be-2c8b9e7ce797

+++ src/com/leclercb/taskunifier/gui/components/configuration/GeneralConfigurationPanel.java
@@ -55,7 +55,7 @@ public class GeneralConfigurationPanel extends ConfigurationPanel {
 		
 		Settings.setLongProperty(
 				"synchronizer.scheduler_sleep_time",
-				((Long) this.getValue("SCHEDULER_SLEEP_TIME")) * 1000l);
+				((Integer) this.getValue("SCHEDULER_SLEEP_TIME")) * 1000l);
 	}
 	
 	private void initialize() {
